requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.25
python-jose==3.3.0
pytest==8.3.3
flake8==7.1.1
bandit==1.7.10
